Updates mostly removing alot of strict notices, also updated all the ajax calls  in the JS files so they use the mootools 1.3  calls, so I think  now for joomla 1.5 you have to have the mootools upgrade plugin to be enabled if that still isn't update.
you can see in the tienda.js file

that I wrapped the Dsc. namespace functions instead the Tienda  versions. 

I did this so it makes the functions use the library, I plan to remove the unused versions form the tienda files over time, also doing that makes it so if anyone built any custom work on for tienda and used those classes we  don't break their code.

 as of right now,  You can add items to your cart, view and update your cart, and  clicking check out fails validation because none of the payment methods  work yet..