a few simple fixes to improving  joomla 2.5  for testing.
Suppressed the errors on Tienda home page, so  it fails nicely. 
Updated Manifest making the submenu module unpublished by default.

fixed the sorting of plugins in shipping view, 

linked Submit Bug Report to the projects  page,  right now it is basically just broke, joomla 2.5 com_admin is drastically different so  migrating it will be  a pretty big job and I feel like it is kinda low on the totem pole.